feat: create useThemeProducts custom hook
